MVC small refactor
