+++ controllers/attraction.py
@@ -3,7 +3,7 @@ import os
 import sys
 PARENT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.insert(0, PARENT_DIR)
-from util import mydb_mgr
+from models import mydb_mgr
 
 app_attraction = Blueprint('app_attraction', __name__)
 mydb = mydb_mgr.mydb_mgr()
